pyproject.toml
[project]
name = "langchain1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.11"
dependencies = [
    "fastapi (==0.115.11)",
    "langchain (==0.3.20)",
    "langchain-community (==0.3.19)",
    "langchain-core (==0.3.41)",
    "langchain-experimental (==0.3.4)",
    "langchain-openai (==0.3.7)",
    "langgraph (==0.3.5)",
    "lxml (==5.3.1)",
    "pandas (==2.2.3)",
    "pydantic (==2.10.6)",
    "pymysql (==1.1.1)",
    "python-dotenv (==1.0.1)",
    "scikit-learn (==1.6.1)",
    "seaborn (==0.13.2)",
    "sqlalchemy (==2.0.38)",
    "sqlglot (==26.8.0)",
    "sqlparse (==0.5.3)",
    "sympy (==1.13.1)",
    "transformers (==4.49.0)",
    "typing-extensions (==4.12.2)",
    "uvicorn (==0.34.0)",
    "weaviate (==0.1.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
